Fix tuple size limit and f64*f32 type mismatch
Split oversized system tuple to stay within Bevy's trait bounds limit.
Cast f32 penalty to f64 for operating_cost multiplication.

+++ crates/simulation/src/water_sources.rs
@@ -493,7 +493,7 @@ mod tests {
         } else {
             1.0
         };
-        let adjusted_cost = base_cost * penalty;
+        let adjusted_cost = base_cost * penalty as f64;
         assert!(
             adjusted_cost > base_cost,
             "Adjusted cost {} should exceed base cost {}",
